+++ src/types/index.ts
@@ -7,11 +7,14 @@ export type { ClickAndVerify, ClickCode } from "./actions";
 export type { CrawlerConfig } from "./config";
 export type {
 	BeforeContext,
+	BlockAutoConfig,
 	BlockContext,
 	BlockHandler,
+	CodeExtractor,
 	PageConfig,
 	PageContext,
 	PageHandler,
+	VariantConfig,
 } from "./handlers";
 export type { CollectionLink, FreeItem, SiteMeta } from "./meta";
 export type { ProgressConfig, ProgressRebuildConfig } from "./progress";